fix lint error

+++ app/helpers/user_signed_in_helper.rb
@@ -8,8 +8,6 @@ module UserSignedInHelper
   end
 
   def user_can_destroy(post)
-    if user_signed_in? && (current_user.id == post.user_id)
-      link_to 'Destroy', post, method: :delete, data: { confirm: 'Are you sure?' }
-    end
+      link_to 'Destroy', post, method: :delete, data: { confirm: 'Are you sure?' } if user_signed_in? && (current_user.id == post.user_id)
   end
 end
